requirements.txt
anyio==4.6.2.post1
boto3==1.35.63
botocore==1.35.63
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jmespath==1.0.1
notion-client==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==1.26.20
